pyproject.toml
[tool.poetry]
name = "twk-backend"
version = "0.1.0"
description = ""
authors = ["Nihar <45964451+nturumel@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "twk_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
cachetools = "^5.3.1"
python-env-loader = "^0.0.2"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
tiktoken = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
sphinx = "^7.0.1"
bandit = "^1.7.5"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.poetry.scripts]
test = "cli:test"
lint = "cli:lint"
typecheck = "cli:typecheck"
format_code = "cli:format_code"
checkformat = "cli:checkformat"
docs = "cli:docs"
launch = "cli:launch"
